+++ WolfoItemBuffs/code/Config.cs
@@ -6,16 +6,22 @@ namespace WolfoItemBuffs
     public class WConfig
     {
         public static ConfigFile ConfigFileUNSORTED = new ConfigFile(Paths.ConfigPath + "\\Wolfo.Wolfo_Item_Buffs.cfg", true);
- 
+
+        public static ConfigEntry<bool> cfg_White_Warbanner;
+
         public static ConfigEntry<bool> cfg_Green_LeechSeed;
         public static ConfigEntry<bool> cfg_Green_Harpoon;
         public static ConfigEntry<float> cfg_Green_Harpoon_VAL;
         public static ConfigEntry<bool> cfg_Green_Whip;
         public static ConfigEntry<bool> cfg_Green_Squid;
+        public static ConfigEntry<bool> cfg_Green_WarCry;
+        public static ConfigEntry<bool> cfg_Green_RegenScrap;
+        public static ConfigEntry<bool> cfg_Green_IgnitionTank;
 
         public static ConfigEntry<bool> cfg_Red_Aegis;
         public static ConfigEntry<float> cfg_Red_Aegis_VAL;
         public static ConfigEntry<bool> cfg_Red_LaserScope;
+        public static ConfigEntry<bool> cfg_Red_ICBM;
 
         public static ConfigEntry<bool> cfg_Yellow_Knurl;
         public static ConfigEntry<bool> cfg_Yellow_DefenseNuc;
@@ -23,13 +29,26 @@ namespace WolfoItemBuffs
         public static ConfigEntry<bool> cfg_Pink_Ring;
 
         public static ConfigEntry<bool> cfg_Blue_Eulogy;
+        public static ConfigEntry<bool> cfg_Blue_Focus;
 
         public static ConfigEntry<bool> cfg_Orange_EliteInherit;
 
 
         public static void InitConfig()
         {
+            cfg_White_Warbanner = ConfigFileUNSORTED.Bind(
+                "White",
+                "Warbanner",
+                true,
+                "Procs on spawn and special bosses."
+            );
             cfg_Green_Harpoon = ConfigFileUNSORTED.Bind(
+                "Green",
+                "Hunters Harpoon",
+                true,
+                "Hunters Harpoon Returns Buffs"
+            );
+            cfg_Green_WarCry = ConfigFileUNSORTED.Bind(
                 "Green",
                 "Hunters Harpoon",
                 true,
@@ -61,8 +80,8 @@ namespace WolfoItemBuffs
             );
             cfg_Red_Aegis_VAL = ConfigFileUNSORTED.Bind(
                 "Red",
-                "Aegis - Value",
-                0.1f,
+                "Aegis - Barrier Decay",
+                0.2f,
                 "Barrier Decay Multiplier"
             );
             //
@@ -70,7 +89,7 @@ namespace WolfoItemBuffs
                 "Red",
                 "Laser Scope",
                 true,
-                "15 crit"
+                "10 crit"
             );
             cfg_Yellow_Knurl = ConfigFileUNSORTED.Bind(
                 "Yellow",
@@ -95,6 +114,12 @@ namespace WolfoItemBuffs
                 "Eulogy Zero",
                 true,
                 "Perfected Elites"
+            );
+            cfg_Blue_Focus = ConfigFileUNSORTED.Bind(
+                "Lunar",
+                "Focused Convergence",
+                true,
+                "Exponential stacking instead linear divided"
             );
             cfg_Orange_EliteInherit = ConfigFileUNSORTED.Bind(
                 "Equipment",